refactor(UserMenu): JS -> TS
