+++ src/base/static/components/molecules/user-menu.tsx
@@ -1,0 +1,299 @@
+/** @jsx jsx */
+import * as React from "react";
+import { jsx, css, InterpolationWithTheme } from "@emotion/core";
+import { useTranslation } from "react-i18next";
+import { connect } from "react-redux";
+import {
+  InternalLink,
+  ExternalLink,
+  SmallText,
+  RegularText,
+} from "../atoms/typography";
+import { Button } from "../atoms/buttons";
+import OfflineDownloadMenu from "../organisms/offline-download-menu";
+import styled from "@emotion/styled";
+import {
+  DashboardsConfig,
+  dashboardConfigSelector,
+} from "../../state/ducks/dashboard-config";
+
+import {
+  userSelector,
+  userPropType,
+  hasAdminAbilities,
+} from "../../state/ducks/user";
+import { OfflineConfig, offlineConfigSelector } from "../../state/ducks/map";
+
+import mq from "../../../../media-queries";
+
+type MenuContainerProps = {
+  isMobileEnabled: boolean;
+};
+const MenuContainer = styled("nav")<MenuContainerProps>(props => ({
+  marginLeft: "auto",
+  marginRight: "10px",
+
+  [mq[0]]: {
+    display: props.isMobileEnabled ? "block" : "none",
+    marginLeft: "auto",
+    marginRight: "auto",
+  },
+  [mq[1]]: {
+    display: "block",
+  },
+}));
+
+const menuStyles = ({
+  isMenuOpen,
+  isLoggedIn,
+}: {
+  isMenuOpen: boolean;
+  isLoggedIn: boolean;
+  // using 'any' for Theme type, until our theme.js is ported over to TS:
+}): InterpolationWithTheme<any> => ({
+  textAlign: "center",
+  float: "left",
+  width: "100%",
+  margin: "0.5em 0",
+  padding: "0",
+  display: isMenuOpen ? "grid" : "none",
+  gridRowGap: "16px",
+  listStyle: "none",
+
+  [mq[1]]: {
+    backgroundColor: "#fff",
+    borderRadius: "3px",
+    boxShadow: "-0.25em 0.25em 0 rgba(0, 0, 0, 0.2)",
+    width: "18em",
+    margin: "0",
+    padding: "1em 0.875em 1.125em 0.875em",
+    position: "absolute",
+    top: "4.125em",
+    right: "1em",
+    // zIndex: "2",  // is z-index valid under a media query?
+
+    "&:before": {
+      content: '""',
+      height: "0",
+      width: "0",
+      border: "1em solid transparent",
+      borderBottomColor: "#666",
+      borderTop: "0",
+      position: "absolute",
+      bottom: "100%",
+      right: isLoggedIn ? "0.5em" : "1.5em",
+    },
+  },
+});
+
+const MenuItem = styled("li")(({ theme }) => ({
+  float: "left",
+  width: "100%",
+  fontFamily: theme.text.navBarFontFamily,
+}));
+
+const SocialMediaMenuItem = styled(MenuItem)({
+  display: "grid",
+  gridTemplateColumns: "1fr 1fr",
+  gridRowGap: "8px",
+  gridColumnGap: "8px",
+});
+
+const SocialLoginButton = styled(ExternalLink)(props => {
+  let backgroundColor;
+  switch (props.service) {
+    case "twitter":
+      backgroundColor = "#4099ff";
+      break;
+    case "facebook":
+      backgroundColor = "#3b5998";
+      break;
+    case "google":
+      backgroundColor = "#e8433a";
+      break;
+  }
+  return {
+    display: "block",
+    padding: "0.5em",
+    boxShadow: props.theme.boxShadow,
+    color: "#fff !important",
+    backgroundColor,
+  };
+});
+
+type StateProps = {
+  currentUser: userPropType;
+
+  hasAdminAbilities: Function;
+  offlineBoundingBox: OfflineConfig;
+  dashboardConfig: DashboardsConfig;
+};
+
+type Props = {
+  pathname: string;
+  apiRoot: string;
+  isMobileEnabled: boolean;
+  isInMobileMode: boolean;
+} & StateProps;
+
+const UserMenu: React.FunctionComponent<Props> = props => {
+  const [isMenuToggled, setIsMenuToggled] = React.useState<boolean>(false);
+  const isMenuOpen = isMenuToggled || props.isInMobileMode;
+
+  const [t, _] = useTranslation(["signIn", "logOut", "signedInAs"]);
+  if (props.currentUser.isAuthenticated) {
+    // If user is logged in
+    const isDashboard = props.pathname === "/dashboard";
+    return (
+      <MenuContainer role="article" isMobileEnabled={props.isMobileEnabled}>
+        {/* TODO: Fix this to use UserAvatar: */}
+        {!props.isInMobileMode && (
+          <img
+            css={{
+              float: "right",
+              width: "2.7em",
+              height: "2.6em",
+              maxWidth: "none",
+              cursor: "pointer",
+
+              fontSize: "1em",
+              padding: "0",
+              outline: "0",
+
+              [mq[0]]: {
+                display: "none",
+              },
+
+              // [mq[1]]: {
+              //   zIndex: "1",  // z-index not valid under media query?
+              // },
+            }}
+            alt="user profile picture"
+            onClick={() => setIsMenuToggled(isMenuOpen => !isMenuOpen)}
+            src={props.currentUser.avatar_url}
+          />
+        )}
+        <ul css={menuStyles({ isMenuOpen, isLoggedIn: true })}>
+          {!!props.dashboardConfig.length &&
+            props.hasAdminAbilities(props.dashboardConfig[0].datasetSlug) && (
+              <MenuItem
+                onClick={() => setIsMenuToggled(isMenuOpen => !isMenuOpen)}
+              >
+                <InternalLink href={isDashboard ? "/" : "/dashboard"}>
+                  {isDashboard ? "back to map" : `go to dashboard`}
+                </InternalLink>
+              </MenuItem>
+            )}
+          {props.offlineBoundingBox && (
+            <MenuItem>
+              <OfflineDownloadMenu
+                offlineBoundingBox={props.offlineBoundingBox}
+              />
+            </MenuItem>
+          )}
+          <MenuItem>
+            <div>
+              <SmallText>{t("signedInAs")}</SmallText> {props.currentUser.name}
+            </div>
+            <ExternalLink
+              css={{
+                fontSize: "0.875em",
+                fontWeight: "normal",
+                textDecoration: "none",
+                textTransform: "uppercase",
+                width: "100%",
+              }}
+              href={`${props.apiRoot}users/logout/`}
+            >
+              {t("logOut")}
+            </ExternalLink>
+          </MenuItem>
+        </ul>
+      </MenuContainer>
+    );
+  } else {
+    // If no user is logged in
+    return (
+      <MenuContainer role="article" isMobileEnabled={props.isMobileEnabled}>
+        {props.isInMobileMode ? (
+          <div
+            css={{
+              width: "100%",
+              display: "flex",
+              justifyContent: "center",
+            }}
+          >
+            <RegularText
+              color="tertiary"
+              height="24px"
+              weight="bold"
+              textTransform="uppercase"
+            >{`Sign In:`}</RegularText>
+          </div>
+        ) : (
+          <Button
+            color="primary"
+            onClick={() => setIsMenuToggled(isMenuOpen => !isMenuOpen)}
+            css={theme => ({
+              fontFamily: theme.text.navBarFontFamily,
+              fontSize: "0.75em",
+              textAlign: "center",
+              textDecoration: "none",
+              lineHeight: "3.25",
+              display: "block",
+              padding: "0 0.5em",
+              height: "100%",
+              cursor: "pointer",
+
+              [mq[0]]: {
+                display: "none",
+              },
+
+              [mq[1]]: {
+                fontSize: "1em",
+                textDecoration: "none",
+                lineHeight: "1.5",
+                padding: "0.5em",
+                position: "relative",
+                zIndex: "3",
+              },
+            })}
+          >
+            {t("signIn")}
+          </Button>
+        )}
+        <ul css={menuStyles({ isMenuOpen, isLoggedIn: false })}>
+          <SocialMediaMenuItem>
+            <SocialLoginButton
+              service={"google"}
+              href={`${props.apiRoot}users/login/google-oauth2/`}
+            >
+              Google
+            </SocialLoginButton>
+            <SocialLoginButton
+              service={"twitter"}
+              href={`${props.apiRoot}users/login/twitter/`}
+            >
+              Twitter
+            </SocialLoginButton>
+            <SocialLoginButton
+              service={"facebook"}
+              href={`${props.apiRoot}users/login/facebook/`}
+            >
+              Facebook
+            </SocialLoginButton>
+          </SocialMediaMenuItem>
+        </ul>
+      </MenuContainer>
+    );
+  }
+};
+
+const mapStateToProps = (state: any): StateProps => ({
+  currentUser: userSelector(state),
+  hasAdminAbilities: datasetSlug => hasAdminAbilities(state, datasetSlug),
+  offlineBoundingBox: offlineConfigSelector(state),
+  dashboardConfig: dashboardConfigSelector(state),
+});
+
+export default connect(mapStateToProps)(UserMenu);
