+++ src/base/static/state/ducks/dashboard-config.js
@@ -1,4 +1,3 @@
-import PropTypes from "prop-types";
 // Selectors:
 export const dashboardConfigSelector = state => {
   return state.dashboardConfig;
@@ -11,20 +10,6 @@ const LOAD = "dashboard-config/LOAD";
 export function loadDashboardConfig(dashboardConfig) {
   return { type: LOAD, payload: dashboardConfig };
 }
-
-// TODO: Remove this when UserMenu is ported to TS, the only place where it's
-// still needed.
-export const dashboardConfigPropType = PropTypes.arrayOf(
-  PropTypes.shape({
-    dastasetId: PropTypes.string,
-    datasetOwner: PropTypes.string,
-    surveyMetrics: PropTypes.shape({
-      categories: PropTypes.bool,
-      demographics: PropTypes.bool,
-      wards: PropTypes.bool,
-    }),
-  }),
-);
 
 const INITIAL_STATE = [];
